Rename snake-case modules to flat names
