+++ Week1Python/Day5MiniProject/DailyChallenge/AdvancedAlgorithm/src/demodata.py
