+++ Week1Python/Day5MiniProject/Exercises/Challenges1/challengessolutions.py
